Move `defaultWorkerFactory.ts` to `/browser/` (#140857)
