+++ src/vs/base/browser/defaultWorkerFactory.ts
