+++ src/vs/editor/browser/services/webWorker.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { URI } from 'vs/base/common/uri';
-import { EditorWorkerClient } from 'vs/editor/common/services/editorWorkerService';
+import { EditorWorkerClient } from 'vs/editor/browser/services/editorWorkerService';
 import { IModelService } from 'vs/editor/common/services/model';
 import * as types from 'vs/base/common/types';
 import { ILanguageConfigurationService } from 'vs/editor/common/languages/languageConfigurationRegistry';
