+++ src/vs/workbench/services/languageDetection/browser/languageDetectionSimpleWorker.ts
@@ -7,13 +7,13 @@ import type { ModelOperations, ModelResult } from '@vscode/vscode-languagedetect
 import { StopWatch } from 'vs/base/common/stopwatch';
 import { IRequestHandler } from 'vs/base/common/worker/simpleWorker';
 import { EditorSimpleWorker } from 'vs/editor/common/services/editorSimpleWorker';
-import { EditorWorkerHost } from 'vs/editor/common/services/editorWorkerService';
+import { IEditorWorkerHost } from 'vs/editor/common/services/editorWorkerHost';
 
 /**
  * Called on the worker side
  * @internal
  */
-export function create(host: EditorWorkerHost): IRequestHandler {
+export function create(host: IEditorWorkerHost): IRequestHandler {
 	return new LanguageDetectionSimpleWorker(host, null);
 }
